WIP: add desktop library
This will be a place to add non-webapp things which can be shared by
desktop wallets. Right now, only the mac app will be able to use this,
but in future there may be other platforms.

This commit also adds generation and confirmation of the wallet recovery
phrase to the desktop library.