Coin contract account tracking
The wallet will now has rudimentary knowledge about coin
contract accounts. The user can add accounts on a certain chain, and see
the balances of those accounts. Deployments will use these accounts to
pay gas instead of the old "sender" field which would pull from the
named keys, which is incorrect. Some stuff had to be rearranged to
facilitate this.

chainwebDefaultSenders will be going away, but it's still useful for us.
It's a little strange at this point in time because they aren't really
full accounts, so you can't get the balance for them.