Add transaction speed slider (#121)
* Add Transaction Speed Slider

This adds the first cut of the transaction speed slider for the deploy
configuration settings.

- probably shouldn't rely on the default gas price as the initial value
- BUG: doesn't display current value on initial load
- Both max and min boundaries are hardcoded

* Update label for transaction speed slider

* Fixes gas price display on load

The gas price is now displayed on modal load. But it is set using the
default gas price value which is hardcoded.

* Updates from code review.

Tighten `mdo` to only the lines that need it.

Inline the `onGasPriceTxt` definition from the let block.

Update the min and max values for the speed slider to never hit zero
and to max out at one. The slider currently doesn't actually hit a
solid value of 1.

* Transaction speed slider uses given gas price

Initial value is incorrect for the moment but the slider scaling
currently works as desired.

* This should fix the display of the value on initial modal load to
be a precision value in the 1e-12 to 1e-8 range.

* Replace gross string functions with actual scaling function.

Needs more testing as I had to use a rounding function and I'm not
sure that I should need to do that.

* Refactor and add type signatures

Moved some of the functions from let to where and gave them type signatures.