Move the data tab to an accordion section.
The data tab from the deploy dialog has been moved to the bottom of the
deploy dialog in a collapsed accordion section.

The previous interface has been replaced with the same interface that is
used on the front IDE panel so they both look and work the same.

The function that creates this interface has been tweaked to take two
functions that are called once a new keyset has been added or after the
JSON input has been successfully updated.