Support signing with capabilities
Adds support for capabilities. The signing API passes all capabilities
it needs, and these are displayed to the user with an account dropdown
box. Other deployment methods allow the user to specify arbitrary
capabilities.

The sender/gas payer field has not been altered here, but in future
could be moved to the capabilities/roles section too.

This also fixes the transaction hash stuff in deployment dialogs. We
were hashing the code, not the command.