Fix progress bar lines.
The lines in the wallet setup header should only change colour as the user
progresses. This fix updates them to be inline elements that update their colour
accordingly.

Updated a slew of styling as well to use the centrally defined attributes from _defs.scss
and cleaned up some of the alignment to be a bit nicer as well.