feat: Adds settings page and removes network cog from toolbars
The idea is that we have some kind of settings page that we can turn
items on/off based on what type off app we are running. Right now, we
only have then network config, and we don't have a non-modal design
so this is just a basic first wave where we:

- Only have network
- Network is still a modal

It's expected that this will need some more work once we have designs for
a non-modal network form and also once we have other settings, but this
feels like a good start.