fix: Better focus handling for passphrase input
Instead of using hover and focus psuedoselectors, this maintains
a Event t Bool for whether the word is focused or not. This is
triggered either by a mouseenter, mouseexit, mousemove or the for element
being tabbed into/out from.

Prior to this, with the psuedoselector the first entry stayed visible
because the user would normally click on that first one to focus and
even once you tab out the mouse is still hovering the element, so it
stays visible.

The best possible fix is to also have an event for when any other word
has been focused to explicitly blur everything else, but that's probably
too fancy and asking for causality loops. :)