fix: Centers the wallet image on the AddAccount success screen
Prior to this we had two images and absolutely positioning one over the other.
This meant moving the base meant twiddling pixels with the other. This now just
uses a block with a background that should be much more robust. The only thing
to be careful of here is the relationship between the height of the div and the
vertical positioning of the inner image, which is still better than before. :)