Updates based on PR feedback
Remove desktop key type as we're never storing it so we don't need the distinction and the
potential for errors was too great.

Add build command function for building a command with a PactKey as a signer.

Revert changes to AccountCreated back to sum type.

Remove accountCreatedBool function.

Tidy up the Receive module to remove unnecessary functions.