Attempt to unify KDA input fields to a single widget.
Add a single purpose widget to be the go to input field when dealing with the KDA
values. This field contains the 'KDA' prefix on the field, the non-negative precision
handling using the coin precision, and also prevents the use of mouse-wheel or up/down
arrows to manipulate the value.