Improve error and request key display for X-Chain Transfers
Errors in crosschain transfers are now displayed in a box that appears below the progress
icons. The box is cleared when the transfer is re-attempted.

The request key no longer displays inline and is display when it is available in a
separate box.

Some effort has been made to pretty-print the error information.