Update UX on add account form
Inflight account details are used to prepopulate the add account form in
the event that the inflight account has errored.

Get balances functionality now uses the 'details' function so there is more
data available for aligning the account updates and ensuring we don't create
accounts with matching names but different keys to ones that exist on the
chain.

Fixed a bug in the `findFirstInflightAccount` function.