Expose nixWrapper and update the README.
Follow literally the instructions in the readme for the NixOS
installation, ie:

```
Run `WEBKIT_DISABLE_COMPOSITING_MODE=1 $(nix-build -A nixosExe)/bin/kadena-chainweaver` to run the linux app in nixos.
```

The resulting app won't be able to connect to testnet and mainnet
because of a problem with certificates. The solution is already present
in the repo, but not exposed to the users; in the linux.nix file, we
read:

```
  nixosWrapper = pkgs.writeScriptBin "${linuxAppName}-wrapper" ''
    #!/usr/bin/env bash
    WEBKIT_DISABLE_COMPOSITING_MODE=1 NIX_SSL_CERT_FILE=${nixosExe}/bin/ca-certificates-patched.crt ${nixosExe}/bin/${linuxAppName}
  '';
```

which fixes bot the webkit problem and the issue with certificates.

This PR exposes the option to build nixosWrapper from the `default.nix`
file, and documents it in the README as the preferred way of building
chainweaver under NixOS.